gumball distrobution
gumble is functioning. If someone wants to take a look at the last plot
it seems kind of strange. just run the gumble section of the main
script.

monthly extrema function needs to be edited to include only data from
the WIS sheet. currently set to export a table using the old format,
and that obviously doesn’t work.